refactor: consolidate client state and hooks
Refs #31
