+++ ROADMAP.md
@@ -126,6 +126,7 @@ This roadmap is the working backlog for turning Anton from a learning harness in
 - [] Add accessible keyboard navigation for settings, sessions, worklog, approvals, and message actions.
 - [✔️] Split chat UI components into feature-oriented folders.
 - [✔️] Split oversized feature components into focused files.
+- [✔️] Consolidate client state and feature API hooks.
 - [] Replace duplicate workspace/settings UI paths with one canonical Settings implementation.
 - [] Clean up layout formatting and UI polish issues that are not core behavior.
